+++ ICD.Connect.Displays.Sony/Properties/AssemblyInfo.cs
@@ -1,0 +1,7 @@
+﻿using System.Reflection;
+
+[assembly: AssemblyTitle("ICD.Connect.Displays.Sony")]
+[assembly: AssemblyCompany("ICD Systems")]
+[assembly: AssemblyProduct("ICD.Connect.Displays.Sony")]
+[assembly: AssemblyCopyright("Copyright © ICD Systems 2018")]
+[assembly: AssemblyVersion("8.0.1.0")]