+++ ICD.Connect.Displays.Sony/SonyBraviaDisplaySettings.cs
@@ -1,0 +1,10 @@
+﻿using ICD.Connect.Displays.Settings;
+using ICD.Connect.Settings.Attributes;
+
+namespace ICD.Connect.Displays.Sony
+{
+	[KrangSettings("SonyBraviaDisplay", typeof(SonyBraviaDisplay))]
+	public sealed class SonyBraviaDisplaySettings : AbstractDisplayWithAudioSettings
+	{
+	}
+}