+++ CHANGELOG.md
@@ -5,9 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+ - Added SonyBraviaDisplay
+
 ### Changed
  - Renamed Microsoft.SurfaceHub project to Microsoft
  - Fixed volume ramping for sharp displays
+ - Only query display state on port assignment if the port is connected
+ - Query displays when port becomes connected
 
 ## [8.0.1] - 2018-09-14
 ### Changed